Restructured Library Hierarchy
Restructured the file hierarchy of the library, moving all currently present neural network related features to a 'nn' module of pyai, for further separation between it and future additions to pyai. Added folders for RL, SL, and searching with module headers and possible ideas to add to each section in future.